notification_service/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
